Working on tile tags

+++ Maker.Hevadea/Game/Storage/LevelStorage.cs
@@ -4,7 +4,7 @@ namespace Maker.Hevadea.Game.Storage
 {
     public class LevelStorage
     {
-        public byte[] Tiles;
+        public int[] Tiles;
         public Dictionary<string, object>[] TilesData;
         public string Name { get; set; }
         public int Id { get; set; }
